+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Added code to handle login acknowledgement.
 - #23 MISSING_JSON_TYPE error code is thrown when attempting to restore the appliance
 - #58 loginMsgAck is not handled in python code
 - #60 SPT type for api1600 should be ServerProfileTemplateV8
+- #67 raise exception when image streamer ip is not provided while running i3s related resources
 
 # 5.2.0
 #### Notes